# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TraderLoading API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /missions/{id}/complete:
    post:
      operationId: completeMission
      tags: [missions]
      summary: Mark a mission as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated mission and profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteMissionResponse"
...
  /ideas:
    get:
      operationId: getIdeas
      tags: [ideas]
      summary: Get ideas and goals
      responses:
        "200":
          description: List of ideas and goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Idea"
    post:
      operationId: createIdea
      tags: [ideas]
      summary: Create an idea or goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIdeaRequest"
      responses:
        "201":
          description: Created idea/goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
  /ideas/{id}:
    put:
      operationId: updateIdea
      tags: [ideas]
      summary: Update an idea or goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIdeaRequest"
      responses:
        "200":
          description: Updated idea/goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
    delete:
      operationId: deleteIdea
      tags: [ideas]
      summary: Delete an idea or goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
  /chat/keys:
    post:
      operationId: savePublicKey
      tags: [chat]
      summary: Save user public key for E2EE
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavePublicKeyBody"
      responses:
        "200":
          description: Key saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicKeyRecord"
  /chat/keys/{userId}:
    get:
      operationId: getPublicKey
      tags: [chat]
      summary: Get a user public key
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicKeyResponse"
        "404":
          description: Key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /chat/messages:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send an encrypted message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessageRecord"
        "403":
          description: Not friends
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /chat/messages/{friendId}:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get messages with a friend
      parameters:
        - name: friendId
          in: path
          required: true
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessagesResponse"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        xp:
          type: integer
        level:
          type: integer
        xpToNextLevel:
          type: integer
        streak:
          type: integer
        levelName:
          type: string
        yearsExperience:
          type: integer
          nullable: true
        winRate:
          type: number
          nullable: true
          description: "Win rate percentage computed from journal entries (0-100)"
        totalTrades:
          type: integer
          nullable: true
      required:
        - id
        - name
        - xp
        - level
        - xpToNextLevel
        - streak
        - levelName
...
    Mission:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        xpReward:
          type: integer
        completed:
          type: boolean
        completedAt:
          type: string
          nullable: true
      required:
        - id
        - title
        - description
        - xpReward
        - completed
...
    CompleteMissionResponse:
      type: object
      properties:
        mission:
          $ref: "#/components/schemas/Mission"
        profile:
          $ref: "#/components/schemas/UserProfile"
        levelUp:
          type: boolean
      required:
        - mission
        - profile
        - levelUp
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Idea:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [idea, goal]
        content:
          type: string
        completed:
          type: boolean
        reminderTime:
          type: string
          nullable: true
        cadence:
          type: string
          enum: [daily, weekly, monthly]
          nullable: true
        recurrence:
          type: boolean
        createdAt:
          type: string
        deadlineDate:
          type: string
          nullable: true
          description: "YYYY-MM-DD deadline date for goals"
        importance:
          type: string
          enum: [low, medium, high]
          nullable: true
      required:
        - id
        - type
        - content
        - completed
        - recurrence
        - createdAt
    CreateIdeaRequest:
      type: object
      properties:
        type:
          type: string
          enum: [idea, goal]
        content:
          type: string
        deadlineDate:
          type: string
          nullable: true
          description: "YYYY-MM-DD deadline date for goals"
        importance:
          type: string
          enum: [low, medium, high]
          nullable: true
      required:
        - type
        - content
    UpdateIdeaRequest:
      type: object
      properties:
        content:
          type: string
        completed:
          type: boolean
        reminderTime:
          type: string
          nullable: true
        cadence:
          type: string
          enum: [daily, weekly, monthly]
          nullable: true
        recurrence:
          type: boolean
        deadlineDate:
          type: string
          nullable: true
          description: "YYYY-MM-DD deadline date for goals"
        importance:
          type: string
          enum: [low, medium, high]
          nullable: true
...
    SavePublicKeyBody:
      type: object
      properties:
        publicKeyJwk:
          type: object
      required:
        - publicKeyJwk
    PublicKeyRecord:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        publicKeyJwk:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - publicKeyJwk
    PublicKeyResponse:
      type: object
      properties:
        userId:
          type: string
        publicKeyJwk:
          type: object
      required:
        - userId
        - publicKeyJwk
    SendMessageBody:
      type: object
      properties:
        receiverId:
          type: string
        ciphertext:
          type: string
        iv:
          type: string
      required:
        - receiverId
        - ciphertext
        - iv
    ChatMessageRecord:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: string
        receiverId:
          type: string
        ciphertext:
          type: string
        iv:
          type: string
        read:
          type: string
        createdAt:
          type: string
      required:
        - id
        - senderId
        - receiverId
        - ciphertext
        - iv
        - createdAt
    ChatMessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessageRecord"
        nextCursor:
          type: integer
          nullable: true
      required:
        - messages
...